Move bulk classification routes into their own module
Keeps create_annotation.py untouched by this PR and gives the bulk
endpoints (and their tests) a dedicated file.
