+++ lightly_studio/src/lightly_studio/api/routes/api/annotation.py
@@ -49,6 +49,7 @@ from lightly_studio.services.annotations_service.update_annotation import (
 
 annotations_router = APIRouter(prefix="/collections/{collection_id}", tags=["annotations"])
 annotations_router.include_router(annotations_module.create_annotation_router)
+annotations_router.include_router(annotations_module.bulk_create_classifications_router)
 annotations_router.include_router(annotations_module.annotation_metrics_info_router)
