+++ lightly_studio/src/lightly_studio/api/routes/api/annotations/__init__.py
@@ -1,11 +1,15 @@
 from .annotation_metrics_info import (
     annotation_metrics_info_router,
 )
+from .bulk_create_classifications import (
+    bulk_create_classifications_router,
+)
 from .create_annotation import (
     create_annotation_router,
 )
 
 __all__ = [
     "annotation_metrics_info_router",
+    "bulk_create_classifications_router",
     "create_annotation_router",
 ]
