+++ lightly_studio/tests/api/routes/api/annotations/test_create_annotation.py
@@ -1,21 +1,12 @@
 from __future__ import annotations
 
-from uuid import UUID, uuid4
+from uuid import UUID
 
 from fastapi.testclient import TestClient
 from pytest_mock import MockerFixture
-from sqlmodel import Session, select
 
-from lightly_studio.api.routes.api.status import (
-    HTTP_STATUS_CREATED,
-    HTTP_STATUS_NOT_FOUND,
-    HTTP_STATUS_OK,
-)
-from lightly_studio.models.annotation.annotation_base import (
-    AnnotationBaseTable,
-    AnnotationType,
-    AnnotationView,
-)
+from lightly_studio.api.routes.api.status import HTTP_STATUS_OK
+from lightly_studio.models.annotation.annotation_base import AnnotationType, AnnotationView
 from lightly_studio.models.annotation.object_detection import ObjectDetectionAnnotationView
 from lightly_studio.models.annotation.segmentation import (
     SegmentationAnnotationView,
@@ -25,7 +16,6 @@ from lightly_studio.models.collection import CollectionTable
 from lightly_studio.models.image import ImageTable
 from lightly_studio.services import annotations_service
 from lightly_studio.services.annotations_service.create_annotation import AnnotationCreateParams
-from tests.helpers_resolvers import ImageStub, create_collection, create_images
 
 
 def test_create_annotation_object_detection(
@@ -256,73 +246,3 @@ def test_create_annotation_with_collection_name(
         created_at=result.created_at,
         tags=[],
     )
-
-
-def test_bulk_create_classifications_route__creates_classifications(
-    db_session: Session,
-    test_client: TestClient,
-) -> None:
-    collection = create_collection(session=db_session)
-    images = create_images(
-        db_session=db_session,
-        collection_id=collection.collection_id,
-        images=[ImageStub(path="a.png"), ImageStub(path="b.png")],
-    )
-
-    response = test_client.post(
-        f"/api/collections/{collection.collection_id}/annotations/classifications/bulk_create",
-        json={
-            "sample_ids": [str(image.sample_id) for image in images],
-            "class_name": "dog",
-            "annotation_collection_name": "ground_truth",
-        },
-    )
-
-    assert response.status_code == HTTP_STATUS_CREATED
-    assert response.json()["created_count"] == 2
-    assert response.json()["skipped_count"] == 0
-    assert {
-        annotation.annotation_type
-        for annotation in db_session.exec(select(AnnotationBaseTable)).all()
-    } == {AnnotationType.CLASSIFICATION}
-
-
-def test_bulk_create_classifications_by_filter_route__skips_existing(
-    db_session: Session,
-    test_client: TestClient,
-) -> None:
-    collection = create_collection(session=db_session)
-    create_images(
-        db_session=db_session,
-        collection_id=collection.collection_id,
-        images=[ImageStub(path="wide.png", width=1920), ImageStub(path="narrow.png", width=10)],
-    )
-    url = (
-        f"/api/collections/{collection.collection_id}"
-        "/annotations/classifications/bulk_create_by_filter"
-    )
-    body = {
-        "filter": {"filter_type": "image", "width": {"min": 100}},
-        "class_name": "dog",
-        "annotation_collection_name": "ground_truth",
-    }
-
-    first = test_client.post(url, json=body)
-    second = test_client.post(url, json=body)
-
-    assert first.status_code == HTTP_STATUS_CREATED
-    assert second.status_code == HTTP_STATUS_CREATED
-    assert first.json()["created_count"] == 1
-    assert second.json()["created_count"] == 0
-    assert second.json()["skipped_count"] == 1
-
-
-def test_bulk_create_classifications_route__unknown_collection(
-    test_client: TestClient,
-) -> None:
-    response = test_client.post(
-        f"/api/collections/{uuid4()}/annotations/classifications/bulk_create",
-        json={"sample_ids": [], "class_name": "dog"},
-    )
-
-    assert response.status_code == HTTP_STATUS_NOT_FOUND
